Add common Dockerfile for all services
The service name can be loaded via env variable so we can use a single
Dockerfile for all services and don't have to handle multiple copies
with very similar code.

This Dockerfile is building a single service in stage 1 and then copying
only the compiled javascript over, installing only production
dependencies and running the entrypoint for each server.