Add timelock for bitcoin
This is running on the testnet only yet..

Implemented the timelock functionality inside the HTLC class.
The class will be expanded to include the hashlock part soon.

It requires an unspent transaction as input (this will come from the
hashlock) as well as the sender and receiver addresses and can be
configured with the amount to send and the time (in blocks) before it
can be refunded.